# specs/001-fastapi-integration/contracts/api-contracts.yaml
openapi: 3.0.0
info:
  title: Robotics Book API
  version: 1.0.0
  description: API for accessing robotics book content
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.robotics-book.com
    description: Production server
paths:
...
  /book/content:
    get:
      summary: Book Content Access
      description: Returns book content in a structured format
      parameters:
        - name: chapter
          in: query
          required: false
          schema:
            type: string
          description: Specific chapter to retrieve
        - name: section
          in: query
          required: false
          schema:
            type: string
          description: Specific section within a chapter
      responses:
        '200':
          description: Book content in structured format
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                  chapter:
                    type: string
                  section:
                    type: string
                  metadata:
                    type: object
        '404':
          description: Content not found
        '500':
          description: Server error